\name{sensorXMLEvent}
\alias{sensorXMLEvent}
\docType{data}
\title{
Sample XML event.
}
\description{
An example XML event, from the Esper documentation.
}
\usage{data(sensorXMLEvent)}
\format{ XML }
\source{
http://esper.codehaus.org
}
\references{
http://esper.codehaus.org
}
\examples{
data(sensorXMLEvent)
# See the example in help('esperr')
}
\keyword{datasets}
